moving util for grouping & sorting entity list to a common folder
